Separate ChangeDetectorTests and ChangePointAggregatorTests.
